fix logic for spot

+++ scripts/ci/run-test-local-env-manager.sh
@@ -122,17 +122,22 @@ function set_instance_env {
     echo pipeline_id: $CI_PIPELINE_ID
     echo job_name: $CI_JOB_NAME
 
-    # export default
-    export X_INFRA_INSTANCE=ondemand
+    if [[ "${FORCED_INFRA_INSTANCE:=""}" != "" ]];then
+	    echo "setting env X_INFRA_INSTANCE to $FORCED_INFRA_INSTANCE"
+      export X_INFRA_INSTANCE=$FORCED_INFRA_INSTANCE
+    else
+      # export default
+      export X_INFRA_INSTANCE=ondemand
 
-    STATUS=$(curl -s -o response.txt -w "%{http_code}" -H "Accept: application/json" "https://gitlab.parity.io/api/v4/projects/${CI_PROJECT_ID}/pipelines/${CI_PIPELINE_ID}/jobs?include_retried=true")
-    if [ $STATUS == "200" ]; then
-      JOBS_CANCELED_BY_US=$(cat response.txt |jq -c ".[] | select((.name | contains(\"${CI_JOB_NAME}\")) and .status == \"canceled\" and (.duration >= 0))") | wc -l
-      if [[ $JOBS_CANCELED_BY_US -eq 0 ]]; then
-        # only mod 10 for PoC
-        R=$(( $CI_PIPELINE_ID % 10));
-        if [[ $R -eq 0 ]];then
-          export X_INFRA_INSTANCE=spot
+      STATUS=$(curl -s -o response.txt -w "%{http_code}" -H "Accept: application/json" "https://gitlab.parity.io/api/v4/projects/${CI_PROJECT_ID}/pipelines/${CI_PIPELINE_ID}/jobs?include_retried=true")
+      if [ $STATUS == "200" ]; then
+        JOBS_CANCELED_BY_US=$(cat response.txt |jq -c ".[] | select((.name | contains(\"${CI_JOB_NAME}\")) and .status == \"canceled\" and (.duration >= 0))" | wc -l)
+        if [[ $JOBS_CANCELED_BY_US -eq 0 ]]; then
+          # only mod 10 for PoC
+          R=$(( $CI_PIPELINE_ID % 10));
+          if [[ $R -eq 0 ]];then
+            export X_INFRA_INSTANCE=spot
+          fi;
         fi;
       fi;
     fi;
